Fixed staticmethod kludge.
Due to the way overrides are handled by ResourceOptions, the
initial_data function appears to be bound to the Resource's Meta
class. If you call that function directly it is expecting Meta
to be passed as 'self'. To get around this I call
initial_data.im_func() instead, which is the unbound function.