Add basic support for inherited django models
It's an kinda ugly idea ugly idea to define a REST-Api around inherited django
models because the resulting API does not look nice but actually tastypie failed
if one tries to do that.

This also adds a fallback to ModelResource._meta.queryset if the attribute
queried in ToManyField.dehydrate does not implement a django manager compatible
api.