Kill the copy, I don't think we really need it.
Do NOT call form.save() - ModelForms do not set defaults based on Models (wtf!). You end up with a Model that has validated, but without the default values. This causes problems with NOT NULL columns in particular - the validation will pass, but the insert will fail. By not calling save here, we let tastypie handle the create and update later.