Patch to get resource lookups working properly when building a related resource.
Now when hydrated a related resource within a payload, we lookup the resource by
its resource_uri if it was provided, instead of provisioning a fresh resource.

The patch that has been implemented feels like it could be done without having
to overload the resource_from_data method. However, in order to get the API
working, this change was required.