Bugfix: Bail out of obj_update if there are no lookup_kwargs.
There was a bug in `obj_update()` that attempted to do a lookup of an existing resource,
despite the incoming resource being new. This results in the lookup_kwargs being empty when passed
to `obj_get()`, and would subsequently return resources that we did not intend to lookup, so any PUTs
would clobber that resource.

This issue has been fixed by raising NotFound if the no lookup_kwargs can be constructed, resulting in a proper
call to `obj_create()`.