Add tests for ModelResource.api_field_from_django_field()
The `api_field_from_django_field()` method of the `ModelResource` class is
mostly untested. The only place that it gets tested is part of
`DecimalFieldTestCase`.

These tests will test the mapping of all API field types that currently have
tests.

**NOTE** The `test_model_resource_correct_association` test of
`DateFieldTestCase` fails. This is the beginning of switching the mapping
from `DateTimeField` to `DateField`. This change will be for #591.