Prepare tests for new field type for DateFields
In order to map Django's `DateField` to Tastypie's, changes need to be made
to the test suite. The values used in conjunction with any field declared as
`DateField` need to be checked against dates rather than datetimes. In order
to not interfere too much with the tests that already exist, I've opted to add
a new field to the `Note` test model.

All relevant tests and fixtures should now account for this additional field.

Again, the test suite should be failing as the mapping hasn't been updated yet.
Plus there is an error in
`core.tests.resources.ModelResourceTestCase.test_get_schema()` related to how
it is using `aware_date()`. Dates do not contain a `tzinfo` property and so an
`AttributeError` is being raised during the test.