Map Django's DateField to Tastypie's DateField
When using a `DateField` in a `Resource` it can be a bit misleading when a datetime is used in the response. This can also lead to validation errors when linking to a `ModelResource`. This introduces the idea of mapping them to `DateField` instead of `DateTimeField`.

As a semantic change, `tastypie.utils.timezone.aware_date()` will now return a date object instead of a datetime.

Unfortunately these changes can potentially be backwards incompatible. This will be noted in the log.