Fixes incorrect assumption that an instance of a model with a non-None primary key has previously been saved to the database.  This is true with the default Django AutoField for the pk, but if you use another field type with a default (e.g. a UUID field), then this test fails.
Added an extra field to Bundle called `obj_is_new` that definitively
indicates whether or not Resource.build_bundle() spawned a new blank
model instance or not.  This allows other logic to conclusively
determine whether or not the fields on bundle.obj should actually be
considered valid.

This requires a slight departure from earlier uses of bundle.obj,
specifically that:

- You should call bundle.save_obj() instead of calling bundle.obj.save()

- You should use bundle.install_new_obj_from_class() instead of just
  assigning a freshly constructed object into the bundle.

- You should use bundle.install_existing_obj() instead of just
  assigning an existing object into the bundle.