Added ``resource_actions`` etc.
Added a very lightweight system to define actions for resources as a
more specific case of ``prepend_urls``

Also moved remove_api_resource_names outside of class, and included two
decorators; ``with_obj`` that wraps around a view function and passses
an extra ``obj`` argument to it when the request specifies enough to
find a object, otherwise it raises exceptions.  Should be used to wrap
``resource_actions`` callbacks. The equivalent for sets,
``with_obj_set`` is also defined.