Add setting to include timezone in serialized ISO datetimes
Add a setting TASTYPIE_DATETIME_FORMATTING_TIMEZONE, to include timezone
in the serialized format of datetimes.  This defaults to False (prior
behaviour).  If the datetime is naive, it is given the default timezone
of the django project.

updated documentation for the added settings.

Add tests to check that the default (False) setting isn't changed,
and that the formatting change happens when it is enabled.

Added timezone related methods to tastypie.utils.timezone:
- make_aware and make_naive: consistant behaviour accross django
  versions, will use get_default_timezone if no tzinfo is passed
- get_default_timezone
- is_naive, is_aware: straight copy from django 1.5

Added tests for make_aware and make_naive to check if it works with the
change in default behaviour.