Return 400 if foreign key matches multiple resources
If a URI matching multiple resources (e.g. the URI corresponding to the
get_list view) were provided as the foreign key of an object, an uncaught
exception would cause a 500 to be returned. MultipleObjectsReturned is now
appropriately handled in resource_from_uri to raise an ApiFieldError, which
results in a 400 being returned.