fixes Django 1.8 regression for saving related fields
Before the django 1.8 upgrade there was a check in resources.py for the
existence of the related obj when saving a bundle. This check was
removed but we still had a few cases where that related obj might be
`None`. This re-adds the condition, and is forwards compatible with
django 1.8.