fix ApiKeyAuthentication for Django 1.8
Django 1.8 checks that the fields given in select_related are correct(https://docs.djangoproject.com/en/1.9/releases/1.8/#select-related-now-checks-given-fields). The select_related method can be used for foreign keys and one to one fields. It is not possible to use it for the reverse relationship User back to ApiKey. so we need to replace select_related with prefetch_related