tests/gis: don't rely on precise object IDs in database
Objects IDs can change, and are generally very brittle. That made the
tests to fail when run in Debian autopkgtest for whatever reason.
Testing coverage isn't actually impacted by relaxing those particular
asserts on the "location", given that the location is used thereafter to
re-query and verify the created object.